Under license control, it now receives files from the server. It's encrypting. I use this file and some data controller. It encrypts and writes to the server again. 1 file has 6 hours life. Even when a new user enters, the file is checked for 6 hours. Do not put unnecessary pressure on the site
Under license control, it now receives files from the server. It's encrypting. I use this file and some data controller. It encrypts and writes to the server again. 1 file has 6 hours life. Even when a new user enters, the file is checked for 6 hours. Do not put unnecessary pressure on the site